+++ lib/src/geofences/geofences.dart
@@ -4,6 +4,7 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:latlong2/latlong.dart';
+import 'package:layrz_logging/layrz_logging.dart';
 import 'package:layrz_models/layrz_models.dart';
 
 part 'geofences.freezed.dart';
@@ -12,3 +13,6 @@ part 'src/category.dart';
 part 'src/geofence.dart';
 part 'src/mode.dart';
 part 'src/point.dart';
+part 'src/variants.dart';
+part 'src/search_item.dart';
+part 'src/export_format.dart';