+++ lib/src/geofences/src/variants.dart
@@ -1,0 +1,12 @@
+part of '../geofences.dart';
+
+enum GeofenceVariant {
+  /// [standard] is the default variant, used in most cases.
+  standard,
+
+  /// [mappit] is the variant used for Mappit geofences.
+  mappit,
+
+  /// [ats] is the variant implemented in the ATS ecosystem.
+  ats,
+}